display image update.

+++ Bsp/inc/bsp_lcd_gui.h
@@ -11,7 +11,7 @@ void TFT_St7789_SetBGColor(uint32_t color);
 
 void TFT_St7789_FillBlock(uint32_t xstart,uint32_t ystart,uint32_t block_width,uint32_t block_height,uint32_t color);
 
-void TFT_ST7789_FillPicture(uint32_t xstart,uint32_t ystart,uint32_t block_width,uint32_t block_height,const uint32_t *black_data);
+void TFT_ST7789_FillPicture(uint32_t xstart,uint32_t ystart,uint32_t block_width,uint8_t block_height,const uint8_t *black_data);
 
 void TFT_ShowChar(uint16_t x,uint16_t y,uint8_t chr,uint8_t fw,uint8_t fh,uint8_t mode);
